+++ heyna.config.js
@@ -7,7 +7,37 @@ module.exports = {
         'check',
         'uncheck',
         'selectOption',
-        'press'
+        'press',
+        'dragAndDrop',
+        'setInputFiles',
+        'hover',
+        'dblclick',
+        'tap',
+        'focus',
+        'blur'
+    ],
+    locatorFactories: [
+        'locator',
+        'getByRole',
+        'getByText',
+        'getByLabel',
+        'getByPlaceholder',
+        'getByTestId',
+        'getByAltText',
+        'getByTitle'
+    ],
+    chainMethods: [
+        'first',
+        'last',
+        'nth',
+        'filter'
+    ],
+    importantActions: [
+        'click',
+        'dblclick',
+        'tap',
+        'setInputFiles',
+        'dragAndDrop'
     ],
     apiLogging: {
         include: [